Report domain, factory and repository

+++ src/main/java/za/ac/cput/repository/ReportRepository.java
@@ -3,5 +3,5 @@ package za.ac.cput.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import za.ac.cput.domain.Report;
 
-public interface ReportRepository extends JpaRepository<Report, Integer> {
+public interface ReportRepository extends JpaRepository<Report, Long> {
 }
